+++ lib/src/stream.dart
@@ -5,19 +5,16 @@ import 'package:meta/meta.dart';
 /// {@template stream.stream_extensions}
 /// Stream extension methods.
 /// {@endtemplate}
-extension StreamX<A> on Stream<A> {
-  /// Allow relieve impact on event loop on large collections.
-  /// Parallelize the event queue and free up time for processing animation,
-  /// user gestures without using isolates.
-  ///
-  /// Thats transformer makes stream low priority.
-  ///
-  /// [duration] - elapsed time of iterations before releasing
-  /// the event queue and microtasks.
+extension BatteriesStreamX<A> on Stream<A> {
+  /// {@macro stream.relieve_stream_transformer}
   Stream<A> relieve([
     Duration duration = const Duration(milliseconds: 4),
   ]) =>
       transform<A>(RelieveStreamTransformer<A>(duration));
+
+  /// {@macro stream.calm_stream_transformer}
+  Stream<A> calm(Duration duration) =>
+      transform<A>(CalmStreamTransformer<A>(duration));
 
   /// {@macro stream.transform_on_type.transformer}
   Stream<A> transformOnType<B extends A>(
@@ -183,5 +180,48 @@ class TransformOnTypeTransformer<A, B extends A>
       ..onCancel = cancelSubscriptions;
 
     return controller.stream;
+  }
+}
+
+/// {@template stream.calm_stream_transformer}
+/// Calm stream transformer.
+/// For example, when you need a pause between events no less than specified.
+/// e.g sending messages to the messenger with intervals and pauses,
+/// in order not to be banned for spam.
+/// {@endtemplate}
+@immutable
+class CalmStreamTransformer<T> extends StreamTransformerBase<T, T> {
+  /// {@macro stream.calm_stream_transformer}
+  const CalmStreamTransformer(this.duration);
+
+  /// Elapsed time of iterations before releasing next event.
+  final Duration duration;
+
+  @override
+  Stream<T> bind(Stream<T> stream) {
+    StreamSubscription<T>? sub;
+    final sc = stream.isBroadcast
+        ? StreamController<T>.broadcast(
+            onCancel: () => sub?.cancel(),
+            sync: false,
+          )
+        : StreamController<T>(
+            onCancel: () => sub?.cancel(),
+            sync: false,
+          );
+    sub = stream.listen(
+      (T value) {
+        sub?.pause();
+        sc.add(value);
+        Future<void>.delayed(duration).then<void>(
+          (_) => sub?.resume(),
+          onError: sc.addError,
+        );
+      },
+      onDone: sc.close,
+      onError: sc.addError,
+      cancelOnError: false,
+    );
+    return sc.stream;
   }
 }