+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.3-dev
+  
+- added: Calm stream transformer
+- **BREAKING**: refactor: Suffix for all extensions
+
 # 0.0.2-dev
   
 - added: Releave extension for Iterable and Stream